requirements.txt
easyocr>=1.7.0
streamlit>=1.28.0
Pillow>=9.0.0
